Fix tests

+++ govcd/catalog_test.go
@@ -133,7 +133,7 @@ func (vcd *TestVCD) Test_DeleteCatalog(check *C) {
 	AddToCleanupList(TestDeleteCatalog, "catalog", vcd.config.VCD.Org, check.TestName())
 	check.Assert(adminCatalog.AdminCatalog.Name, Equals, TestDeleteCatalog)
 
-	checkUploadOvf(vcd, check, vcd.config.OVA.OvaPath, TestDeleteCatalog, TestUploadOvf+"_"+check.TestName())
+	checkUploadOvf(vcd, check, vcd.config.OVA.OvaPath, TestDeleteCatalog, TestUploadOvf+"_"+check.TestName(), false)
 	err = adminCatalog.Delete(false, false)
 	check.Assert(err, NotNil)
 	// Catalog is not empty. An attempt to delete without recursion will fail
@@ -164,7 +164,7 @@ func (vcd *TestVCD) Test_UploadOvf(check *C) {
 	fmt.Printf("Running: %s\n", check.TestName())
 
 	skipWhenOvaPathMissing(vcd.config.OVA.OvaPath, check)
-	checkUploadOvf(vcd, check, vcd.config.OVA.OvaPath, vcd.config.VCD.Catalog.Name, TestUploadOvf)
+	checkUploadOvf(vcd, check, vcd.config.OVA.OvaPath, vcd.config.VCD.Catalog.Name, TestUploadOvf, true)
 }
 
 // Tests System function UploadOvf by creating catalog and
@@ -173,7 +173,7 @@ func (vcd *TestVCD) Test_UploadOvf_chunked(check *C) {
 	fmt.Printf("Running: %s\n", check.TestName())
 
 	skipWhenOvaPathMissing(vcd.config.OVA.OvaChunkedPath, check)
-	checkUploadOvf(vcd, check, vcd.config.OVA.OvaChunkedPath, vcd.config.VCD.Catalog.Name, TestUploadOvf+"2")
+	checkUploadOvf(vcd, check, vcd.config.OVA.OvaChunkedPath, vcd.config.VCD.Catalog.Name, TestUploadOvf+"2", true)
 }
 
 // Tests System function UploadOvf by creating catalog and
@@ -309,7 +309,7 @@ func (vcd *TestVCD) Test_UploadOvfFile(check *C) {
 	fmt.Printf("Running: %s\n", check.TestName())
 
 	skipWhenOvaPathMissing(vcd.config.OVA.OvfPath, check)
-	checkUploadOvf(vcd, check, vcd.config.OVA.OvfPath, vcd.config.VCD.Catalog.Name, TestUploadOvf+"7")
+	checkUploadOvf(vcd, check, vcd.config.OVA.OvfPath, vcd.config.VCD.Catalog.Name, TestUploadOvf+"7", true)
 }
 
 // Tests System function UploadOvf by creating catalog and
@@ -318,7 +318,7 @@ func (vcd *TestVCD) Test_UploadOvf_withoutVMDKSize(check *C) {
 	fmt.Printf("Running: %s\n", check.TestName())
 
 	skipWhenOvaPathMissing(vcd.config.OVA.OvaWithoutSizePath, check)
-	checkUploadOvf(vcd, check, vcd.config.OVA.OvaWithoutSizePath, vcd.config.VCD.Catalog.Name, TestUploadOvf+"8")
+	checkUploadOvf(vcd, check, vcd.config.OVA.OvaWithoutSizePath, vcd.config.VCD.Catalog.Name, TestUploadOvf+"8", true)
 }
 
 func countFolders() int {
@@ -335,7 +335,7 @@ func countFolders() int {
 	return count
 }
 
-func checkUploadOvf(vcd *TestVCD, check *C, ovaFileName, catalogName, itemName string) {
+func checkUploadOvf(vcd *TestVCD, check *C, ovaFileName, catalogName, itemName string, deleteItemAtTheEnd bool) {
 	catalog, org := findCatalog(vcd, check, catalogName)
 
 	uploadTask, err := catalog.UploadOvf(ovaFileName, itemName, "upload from test", 1024)
@@ -350,7 +350,9 @@ func checkUploadOvf(vcd *TestVCD, check *C, ovaFileName, catalogName, itemName s
 	verifyCatalogItemUploaded(check, catalog, itemName)
 
 	// Delete testing catalog item
-	deleteCatalogItem(check, catalog, itemName)
+	if deleteItemAtTheEnd {
+		deleteCatalogItem(check, catalog, itemName)
+	}
 }
 
 func verifyCatalogItemUploaded(check *C, catalog *Catalog, itemName string) {
@@ -387,7 +389,7 @@ func skipWhenOvaPathMissing(ovaPath string, check *C) {
 }
 
 func deleteCatalogItem(check *C, catalog *Catalog, itemName string) {
-	catalogItem, err := catalog.GetCatalogItemByName(itemName, false)
+	catalogItem, err := catalog.GetCatalogItemByName(itemName, true)
 	check.Assert(err, IsNil)
 	check.Assert(catalogItem, NotNil)
